context/context-api/test_weaviate_connection.py
#!/usr/bin/env python3
"""
Weaviate Cloud Connection Test for Context API Service
======================================================

Test script to validate Weaviate Cloud connectivity and functionality
for the context-api service migration from Qdrant.

Features:
- Validates connection to Weaviate Cloud instance
- Tests basic schema operations
- Verifies document storage and retrieval
- Tests semantic search capabilities
- Provides detailed connection diagnostics

Usage:
    python test_weaviate_connection.py

Environment Variables Required:
    WEAVIATE_URL - Weaviate Cloud endpoint URL
    WEAVIATE_API_KEY - Authentication key for Weaviate Cloud

Author: Sophia AI Intelligence Team
Version: 1.0.0 (Context API Migration)
"""

import os
import sys
import time
import uuid
from datetime import datetime, timezone

import weaviate

# Weaviate Cloud connection configuration
WEAVIATE_URL = os.getenv("WEAVIATE_URL", "https://w6bigpoxsrwvq7wlgmmdva.c0.us-west3.gcp.weaviate.cloud")
WEAVIATE_API_KEY = os.getenv("WEAVIATE_API_KEY", "VMKjGMQUnXQIDiFOciZZOhr7amBfCHMh7hNf")

def test_weaviate_connection():
    """
    Comprehensive test of Weaviate Cloud connection and functionality.
    
    Tests performed:
    1. Basic connectivity and authentication
    2. Schema management and class creation
    3. Document object creation and storage
    4. Semantic search with vector queries
    5. Data cleanup and schema validation
    
    Returns:
        bool: True if all tests pass, False otherwise
    """
    print("🧪 WEAVIATE CLOUD CONNECTION TEST - CONTEXT API SERVICE")
    print("=" * 60)
    print(f"🌐 Testing endpoint: {WEAVIATE_URL}")
    print(f"🔑 API Key configured: {'✅ Yes' if WEAVIATE_API_KEY else '❌ No'}")
    print("")

    if not WEAVIATE_URL or not WEAVIATE_API_KEY:
        print("❌ ERROR: Missing required environment variables")
        print("   Required: WEAVIATE_URL, WEAVIATE_API_KEY")
        return False

    try:
        print("📡 Step 1: Establishing Weaviate Cloud connection...")
        client = weaviate.Client(
            url=WEAVIATE_URL,
            auth_client_secret=weaviate.AuthApiKey(WEAVIATE_API_KEY)
        )
        
        # Test basic connectivity
        if not client.is_ready():
            print("❌ Connection failed: Weaviate instance not ready")
            return False
            
        print("✅ Connection established successfully")

        print("\n📊 Step 2: Checking cluster status...")
        try:
            cluster_status = client.cluster.get_nodes_status()
            print(f"✅ Cluster status: {len(cluster_status)} nodes online")
            for i, node in enumerate(cluster_status):
                print(f"   Node {i+1}: {node.get('status', 'unknown')} - {node.get('name', 'unnamed')}")
        except Exception as e:
            print(f"⚠️  Cluster status check failed: {e}")

        print("\n🏗️  Step 3: Setting up test schema...")
        
        # Define Documents schema for context API
        documents_class = {
            "class": "ContextDocuments",
            "description": "Context API document storage for testing",
            "properties": [
                {
                    "name": "content",
                    "dataType": ["text"],
                    "description": "Document content"
                },
                {
                    "name": "source",
                    "dataType": ["string"], 
                    "description": "Document source or URL"
                },
                {
                    "name": "account_id",
                    "dataType": ["string"],
                    "description": "Account identifier"
                },
                {
                    "name": "doc_id",
                    "dataType": ["string"],
                    "description": "Document unique identifier"
                },
                {
                    "name": "metadata",
                    "dataType": ["text"],
                    "description": "Additional document metadata (JSON string)"
                },
                {
                    "name": "created_at",
                    "dataType": ["string"],
                    "description": "Creation timestamp"
                }
            ]
        }

        # Clean up any existing test class
        try:
            client.schema.delete_class("ContextDocuments")
            print("🧹 Cleaned up existing test schema")
        except:
            pass  # Class doesn't exist, that's fine

        # Create the class
        client.schema.create_class(documents_class)
        print("✅ Test schema 'ContextDocuments' created")

        print("\n📝 Step 4: Testing document storage...")
        
        # Test document data
        test_doc_id = str(uuid.uuid4())
        test_document = {
            "content": "This is a test document for the Context API service migration. It contains sample content to verify vector storage and semantic search capabilities in Weaviate Cloud.",
            "source": "test_migration_document",
            "account_id": "test_account_001",
            "doc_id": test_doc_id,
            "metadata": '{"type": "test", "migration": "qdrant_to_weaviate", "service": "context-api"}',
            "created_at": datetime.now(timezone.utc).isoformat()
        }

        # Store document with vector embedding
        client.data_object.create(
            data_object=test_document,
            class_name="ContextDocuments",
            uuid=test_doc_id
        )
        
        print(f"✅ Test document stored with ID: {test_doc_id}")

        # Wait for indexing
        print("⏳ Waiting for document indexing...")
        time.sleep(3)

        print("\n🔍 Step 5: Testing semantic search...")
        
        # Test semantic search
        search_query = "test migration document"
        result = (
            client.query
            .get("ContextDocuments", ["content", "source", "account_id", "doc_id", "metadata"])
            .with_near_text({"concepts": [search_query]})
            .with_limit(1)
            .with_additional(["certainty", "id"])
            .do()
        )

        if "data" in result and "Get" in result["data"] and "ContextDocuments" in result["data"]["Get"]:
            documents = result["data"]["Get"]["ContextDocuments"]
            if documents:
                doc = documents[0]
                certainty = doc["_additional"]["certainty"]
                print(f"✅ Search successful - found document with certainty: {certainty:.3f}")
                print(f"   Content preview: {doc['content'][:100]}...")
                print(f"   Source: {doc['source']}")
                print(f"   Account ID: {doc['account_id']}")
            else:
                print("⚠️  Search returned no results")
        else:
            print("❌ Search query failed")

        print("\n📈 Step 6: Testing aggregation queries...")
        
        # Test object count
        try:
            count_result = client.query.aggregate("ContextDocuments").with_meta_count().do()
            doc_count = count_result.get("data", {}).get("Aggregate", {}).get("ContextDocuments", [{}])[0].get("meta", {}).get("count", 0)
            print(f"✅ Document count query successful: {doc_count} documents")
        except Exception as e:
            print(f"⚠️  Count query failed: {e}")

        print("\n🧹 Step 7: Cleaning up test data...")
        
        # Clean up test document
        try:
            client.data_object.delete(
                uuid=test_doc_id,
                class_name="ContextDocuments"
            )
            print("✅ Test document deleted")
        except Exception as e:
            print(f"⚠️  Document cleanup failed: {e}")

        # Clean up test schema
        try:
            client.schema.delete_class("ContextDocuments")
            print("✅ Test schema deleted")
        except Exception as e:
            print(f"⚠️  Schema cleanup failed: {e}")

        print("\n" + "=" * 60)
        print("🎉 ALL TESTS PASSED - CONTEXT API WEAVIATE CONNECTION VERIFIED")
        print("✅ Connection: Ready")
        print("✅ Authentication: Valid")
        print("✅ Schema operations: Working")
        print("✅ Document storage: Working")
        print("✅ Semantic search: Working") 
        print("✅ Data cleanup: Working")
        print("")
        print("🚀 Context API service is ready for Weaviate Cloud migration!")
        
        # Note: Weaviate v3 client doesn't have close() method
        
        return True

    except weaviate.exceptions.AuthenticationFailedException:
        print("\n❌ AUTHENTICATION FAILED")
        print("   Check your WEAVIATE_API_KEY")
        print("   Ensure the API key is valid and has proper permissions")
        return False
        
    except weaviate.exceptions.UnexpectedStatusCodeException as e:
        print(f"\n❌ WEAVIATE API ERROR: {e}")
        print("   Check your WEAVIATE_URL")
        print("   Verify the endpoint is accessible")
        return False
        
    except Exception as e:
        print(f"\n❌ UNEXPECTED ERROR: {e}")
        print("   Full error details:")
        import traceback
        traceback.print_exc()
        return False

def main():
    """Main test execution"""
    success = test_weaviate_connection()
    
    if success:
        print("\n🔧 MIGRATION READINESS:")
        print("   ✅ Context API can connect to Weaviate Cloud")
        print("   ✅ Vector operations are functional")
        print("   ✅ Search capabilities are working")
        print("   ✅ Ready to replace Qdrant configuration")
        sys.exit(0)
    else:
        print("\n🚨 MIGRATION BLOCKED:")
        print("   ❌ Weaviate Cloud connection issues detected")
        print("   ❌ Fix connection problems before proceeding")
        print("   ❌ Context API migration cannot proceed")
        sys.exit(1)

if __name__ == "__main__":
    main()