context/context-api/requirements.txt
# Sophia AI Context API Service Dependencies
# =========================================

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database connectivity
asyncpg==0.29.0

# Vector database
weaviate-client==3.24.2

# HTTP and async support
httpx==0.25.2
aiohttp==3.9.1

# Utility libraries
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Logging and monitoring
structlog==23.2.0